// src/models/Product.d.ts
import { Product } from '../types';
export declare class ProductItem implements Product {
    id: number;
    title: string;
    description: string;
    price: number;
    discountPercentage: number;
    rating: number;
    stock: number;
    brand: string;
    category: string;
    thumbnail: string;
    images: string[];
    constructor(data: Product);
}